refactor(server): #8 split monolithic index.js into config/, routes/, and websocket.js
All 18 API endpoints preserved with identical behaviour. Routes distributed
across auth, orders, products, admin, coffee, and qr modules. CORS and
multer config extracted to config/, WebSocket init and broadcast helpers
to websocket.js.
