+++ server/config/cors.js
@@ -1,0 +1,17 @@
+const corsOptions = {
+  origin: process.env.ALLOWED_ORIGINS,
+  methods: ['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS'],
+  allowedHeaders: [
+    'Content-Type',
+    'Authorization',
+    'Upgrade',
+    'Connection',
+    'Sec-WebSocket-Key',
+    'Sec-WebSocket-Version',
+    'Sec-WebSocket-Protocol',
+    'Sec-WebSocket-Extensions'
+  ],
+  credentials: true
+};
+
+module.exports = corsOptions;
